+++ crates/net/headers-downloaders/Cargo.toml
@@ -17,6 +17,6 @@ reth-rpc-types = { path = "../rpc-types" }
 assert_matches = "1.5.0"
 once_cell = "1.15.0"
 rand = "0.8.5"
-reth-interfaces = { path = "../../interfaces", features = ["test-helpers"] }
+reth-interfaces = { path = "../../interfaces", features = ["test-utils"] }
 tokio = { version = "1.21.2", features = ["full"] }
 serial_test = "0.9.0"
